pyproject.toml
[project]
name = "general-agent"
version = "0.1.0"
description = "通用Agent系统，兼容Skill协议、MCP和RAG"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.3",
    "aiosqlite>=0.19.0",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "mcp>=0.9.0",
]

[project.scripts]
agent = "src.cli.__main__:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
]
rag = [
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.0",
    "pypdf>=3.17.0",
    "tiktoken>=0.5.0",
    "markdown>=3.5.0",
]
cli = [
    "textual>=0.47.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
